requirements.txt
Flask==2.1.0
Werkzeug==2.1.0
python-dotenv==0.19.0
